backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
python-multipart==0.0.20
pydantic>=2.0.0
pydantic-settings>=2.0.0
scikit-learn>=1.5.0
pytest>=8.0.0
httpx>=0.27.0

# Voice-to-RAG (음성→지식베이스)
pydub>=0.25.1
ffmpeg-python>=0.2.0
# openai-whisper  # (선택) 로컬 Whisper STT — GPU 권장, 설치: pip install openai-whisper

# 웹 크롤러 (홈페이지→지식베이스)
beautifulsoup4>=4.12.0
lxml>=5.0.0

# 데이터 허브 — HuggingFace 데이터셋 다운로드/처리
huggingface_hub>=0.20.0
datasets>=2.16.0
requests>=2.31.0
